requirements.txt
pandas
fyers-apiv2
SQLAlchemy
ta